remove one level of indirection with ProcessedRequest, sign after persisting
